+++ testutil/network.go
@@ -55,7 +55,7 @@ func DefaultTestNetworkConfig() testnet.Config {
 		log.NewNopLogger(), dbm.NewMemDB(), nil, true, make(map[int64]bool), tempDir, 0,
 		simtestutil.NewAppOptionsWithFlagHome(tempDir),
 	)
-	encCfg := provenanceapp.MakeTestEncodingConfig(nil)
+	encCfg := tempApp.GetEncodingConfig()
 
 	return testnet.Config{
 		Codec:             encCfg.Marshaler,
@@ -94,8 +94,6 @@ func Cleanup(n *testnet.Network, t *testing.T) {
 	}
 	t.Log("Cleanup: Cleaning up testnet.")
 	n.Cleanup()
-	// Give things a chance to finish closing up. Hopefully will prevent things like address collisions. 100ms chosen randomly.
-	time.Sleep(100 * time.Millisecond)
 	t.Log("Cleanup: Done.")
 }