+++ x/msgfees/client/cli/cli_test.go
@@ -12,7 +12,6 @@ import (
 	cmtcli "github.com/cometbft/cometbft/libs/cli"
 
 	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/crypto/hd"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	testnet "github.com/cosmos/cosmos-sdk/testutil/network"
@@ -34,17 +33,16 @@ type IntegrationTestSuite struct {
 	cfg     testnet.Config
 	testnet *testnet.Network
 
+	keyring          keyring.Keyring
+	keyringEntries   []testutil.TestKeyringEntry
+	accountAddresses []sdk.AccAddress
+
 	accountAddr sdk.AccAddress
 	accountKey  *secp256k1.PrivKey
 
 	account2Addr  sdk.AccAddress
 	account2Key   *secp256k1.PrivKey
 	acc2NameCount int
-
-	accountAddresses []sdk.AccAddress
-
-	keyring    keyring.Keyring
-	keyringDir string
 }
 
 func TestIntegrationTestSuite(t *testing.T) {
@@ -107,25 +105,14 @@ func (s *IntegrationTestSuite) SetupSuite() {
 	s.testnet, err = testnet.New(s.T(), s.T().TempDir(), s.cfg)
 	s.Require().NoError(err, "creating testnet")
 
-	s.testnet.Validators[0].ClientCtx = s.testnet.Validators[0].ClientCtx.WithKeyringDir(s.keyringDir).WithKeyring(s.keyring)
+	s.testnet.Validators[0].ClientCtx = s.testnet.Validators[0].ClientCtx.WithKeyring(s.keyring)
 	_, err = testutil.WaitForHeight(s.testnet, 1)
 	s.Require().NoError(err, "waiting for height 1")
 }
 
 func (s *IntegrationTestSuite) GenerateAccountsWithKeyrings(number int) {
-	path := hd.CreateHDPath(118, 0, 0).String()
-	s.keyringDir = s.T().TempDir()
-	kr, err := keyring.New(s.T().Name(), "test", s.keyringDir, nil, s.cfg.Codec)
-	s.Require().NoError(err, "Keyring.New")
-	s.keyring = kr
-	for i := 0; i < number; i++ {
-		keyId := fmt.Sprintf("test_key%v", i)
-		info, _, err := kr.NewMnemonic(keyId, keyring.English, path, keyring.DefaultBIP39Passphrase, hd.Secp256k1)
-		s.Require().NoError(err, "Keyring.NewMnemonic")
-		addr, err := info.GetAddress()
-		s.Require().NoError(err, "GetAddress")
-		s.accountAddresses = append(s.accountAddresses, addr)
-	}
+	s.keyringEntries, s.keyring = testutil.GenerateTestKeyring(s.T(), number, s.cfg.Codec)
+	s.accountAddresses = testutil.GetKeyringEntryAddresses(s.keyringEntries)
 }
 
 func (s *IntegrationTestSuite) TearDownSuite() {